requirements.txt
# Core LLM & Framework
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-cohere>=0.3.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0

# Embeddings & Vector Store
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Sparse Retrieval
rank-bm25>=0.2.2

# MCP Server
mcp>=1.0.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.0

# Visualization
plotly>=5.22.0
networkx>=3.3

# Evaluation
ragas>=0.2.0

# PDF Processing
pypdf>=4.0.0

# Web UI
streamlit>=1.37.0

# Environment
python-dotenv>=1.0.0
pydantic>=2.7.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
